refactor(sglang): colocate sidecar Python entrypoint
