+++ lib/bindings/python/rust/lib.rs
@@ -82,7 +82,6 @@ mod parsers;
 mod planner;
 mod prometheus_metrics;
 mod python_payload;
-mod sglang_sidecar;
 
 type PythonServerStreamingIngress = Ingress<
     SingleIn<python_payload::PythonPayload>,
@@ -177,7 +176,7 @@ fn _core(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(run_kv_indexer, m)?)?;
     m.add_function(wrap_pyfunction!(run_slot_tracker, m)?)?;
     m.add_function(wrap_pyfunction!(run_select_service, m)?)?;
-    m.add_function(wrap_pyfunction!(sglang_sidecar::run_sglang_sidecar, m)?)?;
+    m.add_function(wrap_pyfunction!(llm::entrypoint::run_sglang_sidecar, m)?)?;
     m.add_function(wrap_pyfunction!(llm::entrypoint::make_engine, m)?)?;
     m.add_function(wrap_pyfunction!(llm::replay::run_mocker_trace_replay, m)?)?;
     m.add_function(wrap_pyfunction!(
